Improve expected values estimation
The new strategy involves two new steps:
- Dynamic bin aggregation
- Outlier detection and correction

The rationale behind the dynamic bin aggregation is the following:
- The number of possible and observed data points used to estimate the
  tail part of the expected value profile is often fairly low, leading
  to noisy and unreliable estimations.
- To deal with this issue, we aggregate bins with fewer than N_obs
  observed_values and fewer than N_pos possible_values (both parameters
  can be tuned at the API and CLI level)
- In case the last bin fails to meet the above cutoffs, then the
  expected value for the (possibly aggregated) bin is set to NaN

From our testing, the above approach better estimates the expected
values for sparse matrices.

That being said, there are other factors that can make the expected
curve not monotonically decreasing with genomic distance and/or not
smooth, such as outliers.

From our testing, outliers are usually due to translocation events that
are too small to be masked by the MAD-max filter. This can result in
sharp peaks along the expected value curve.

We deal with these outliers (regardless of their origin as follows):
- Loop over the expected value vector using a window size of w bp
- For each window w_i, compute e.g. the 0.975 quantile, if weight w[i]
  is greater than the quantile, flag the weight as an outlier
- Outliers are re-estimated by computing the median value over the
  window w_i, and replacing the original weight with the median value,
  but only if the median is a finite number

Both the quantile and the window size can be tweaked from the API and
CLI.

Unfortunately, it is unlikely that there is a combination of quantile
and window size that will work for all Hi-C matrices, so the default
parameters are set to deal with outliers spanning 250kbp or less.

Users are advised to visualize the expected value profile computed by
NCHG expected, and increase the window size and/or decrease the quantile
value until most of the outliers are detected and correctly re-estimated.