Add a build knob to disable the use of pidfd_open()
This is mostly needed to ensure that the binaries produced by compiling
NCHG on Linux systems running kernel 5.14 or newer are able to run
also on systems with older kernels.
This is especially important for the containerized version of NCHG, as
the Docker images are usually built using cloud/dev machines running
recent versions of the Linux kernel, while the image itself could be
deployed on systems running dated versions of the Linux kernel.